Tags, Search and Filter
User can now search for a product based on a search  query, a category or based on tags.
Created Tags model
Created tags and box_tag migration.
Created a factory for tags.
Updated the Seeder to create tags, boxes and assign tags at random to each box.
Updated the box index to add a search bar and a filters drop down. Also moved the categories to the filter drop down.
Updated the BoxController to handle search logic based on query, type, tags.

Added fontawsome to the project for better icons.
Updated nav bar to  use fontawsome icon and fixed bug.